Move ui/index_patterns files to data plugin.
