+++ src/legacy/core_plugins/data/public/index_patterns/components/index_pattern_select.tsx
