+++ src/legacy/core_plugins/data/public/index_patterns/errors.ts
