+++ src/legacy/core_plugins/data/public/index_patterns/fields/fields_fetcher.ts
@@ -17,32 +17,30 @@
  * under the License.
  */
 
-// @ts-ignore
-import { get } from 'lodash';
-// @ts-ignore
-import { KBN_FIELD_TYPES } from '../../../../utils/kbn_field_types';
-import { Field } from '../_field';
+import { GetFieldsOptions, IndexPattern, IndexPatternsApiClient } from '../index_patterns';
 
-const filterableTypes = KBN_FIELD_TYPES.filter((type: any) => type.filterable).map(
-  (type: any) => type.name
-);
-
-export function isFilterable(field: Field): boolean {
-  return (
-    field.name === '_id' ||
-    field.scripted ||
-    (field.searchable && filterableTypes.includes(field.type))
-  );
-}
-
-export function getFromSavedObject(savedObject: any) {
-  if (get(savedObject, 'attributes.fields') === undefined) {
-    return;
-  }
-
-  return {
-    id: savedObject.id,
-    fields: JSON.parse(savedObject.attributes.fields),
-    title: savedObject.attributes.title,
+export const createFieldsFetcher = (
+  indexPattern: IndexPattern,
+  apiClient: IndexPatternsApiClient,
+  metaFields: string
+) => {
+  const fieldFetcher = {
+    fetch: (options: GetFieldsOptions) => {
+      return fieldFetcher.fetchForWildcard(indexPattern.title, {
+        ...options,
+        type: indexPattern.type,
+        params: indexPattern.typeMeta && indexPattern.typeMeta.params,
+      });
+    },
+    fetchForWildcard: (pattern: string, options: GetFieldsOptions = {}) => {
+      return apiClient.getFieldsForWildcard({
+        pattern,
+        metaFields,
+        type: options.type,
+        params: options.params || {},
+      });
+    },
   };
-}
+
+  return fieldFetcher;
+};
