+++ src/legacy/core_plugins/data/public/index_patterns/index_patterns/_flatten_hit.js
