+++ src/legacy/core_plugins/data/public/index_patterns/index_patterns/index_patterns.ts
@@ -24,10 +24,10 @@ import {
   UiSettingsClientContract,
 } from 'src/core/public';
 // @ts-ignore
-import { fieldFormats } from '../registry/field_formats';
+import { fieldFormats } from 'ui/registry/field_formats';
 
-import { IndexPattern } from './_index_pattern';
 import { createIndexPatternCache } from './_pattern_cache';
+import { IndexPattern } from './index_pattern';
 import { IndexPatternsApiClient } from './index_patterns_api_client';
 const indexPatternCache = createIndexPatternCache();
 const apiClient = new IndexPatternsApiClient();
