+++ src/legacy/core_plugins/data/public/index_patterns/utils.test.ts
@@ -17,28 +17,53 @@
  * under the License.
  */
 
-import expect from '@kbn/expect';
-import { isFilterable } from '../index';
+import {
+  CONTAINS_SPACES,
+  ILLEGAL_CHARACTERS,
+  INDEX_PATTERN_ILLEGAL_CHARACTERS_VISIBLE,
+  isFilterable,
+  validateIndexPattern,
+} from './utils';
+
+import { Field } from './fields';
 
 const mockField = {
   name: 'foo',
   scripted: false,
   searchable: true,
   type: 'string',
-};
+} as Field;
 
-describe('static utils', () => {
+describe('Index Pattern Utils', () => {
+  describe('Validation', () => {
+    it('should not allow space in the pattern', () => {
+      const errors = validateIndexPattern('my pattern');
+      expect(errors[CONTAINS_SPACES]).toBe(true);
+    });
+
+    it('should not allow illegal characters', () => {
+      INDEX_PATTERN_ILLEGAL_CHARACTERS_VISIBLE.forEach(char => {
+        const errors = validateIndexPattern(`pattern${char}`);
+        expect(errors[ILLEGAL_CHARACTERS]).toEqual([char]);
+      });
+    });
+
+    it('should return empty object when there are no errors', () => {
+      expect(validateIndexPattern('my-pattern-*')).toEqual({});
+    });
+  });
+
   describe('isFilterable', () => {
     describe('types', () => {
       it('should return true for filterable types', () => {
         ['string', 'number', 'date', 'ip', 'boolean'].forEach(type => {
-          expect(isFilterable({ ...mockField, type })).to.be(true);
+          expect(isFilterable({ ...mockField, type })).toBe(true);
         });
       });
 
       it('should return false for filterable types if the field is not searchable', () => {
         ['string', 'number', 'date', 'ip', 'boolean'].forEach(type => {
-          expect(isFilterable({ ...mockField, type, searchable: false })).to.be(false);
+          expect(isFilterable({ ...mockField, type, searchable: false })).toBe(false);
         });
       });
 
@@ -52,18 +77,17 @@ describe('static utils', () => {
           'unknown',
           'conflict',
         ].forEach(type => {
-          expect(isFilterable({ ...mockField, type })).to.be(false);
+          expect(isFilterable({ ...mockField, type })).toBe(false);
         });
       });
     });
 
-
     it('should return true for scripted fields', () => {
-      expect(isFilterable({ ...mockField, scripted: true, searchable: false })).to.be(true);
+      expect(isFilterable({ ...mockField, scripted: true, searchable: false })).toBe(true);
     });
 
     it('should return true for the _id field', () => {
-      expect(isFilterable({ ...mockField, name: '_id' })).to.be(true);
+      expect(isFilterable({ ...mockField, name: '_id' })).toBe(true);
     });
   });
 });
